fix: security hardening, cppcheck fixes, explicit ctors, CI lint/test/cppcheck, translate Italian strings
test: add Catch2 unit tests for array, TLV, crypto, ASN.1

- Add catch2 wrap (v3.14.0) and tests/ build target
- Tests cover ByteArray/ByteDynArray, TLV parser/builder,
  SHA1/SHA256/SHA512, Base64, and ASN.1 parser (47 cases)
- Fix base64.cpp: replace broken ArraySink with CryptoPP
  filter-chain pattern (StringSink owned by StringSource)

+++ pkcs11/src/sign/cie_sign.cpp
@@ -28,8 +28,8 @@ uint16_t CIESign::sign(const char* inFilePath, const char* type,
     if (ret != 0) {
       throw ret;
     }
-    ret =
-        cie_sign_sign_set(ctx, CIE_SIGN_OPT_CADES, reinterpret_cast<void*>(1));
+    ret = cie_sign_sign_set(ctx, CIE_SIGN_OPT_CADES,
+                            reinterpret_cast<void*>(static_cast<uintptr_t>(1)));
     if (ret != 0) {
       throw ret;
     }
@@ -53,7 +53,8 @@ uint16_t CIESign::sign(const char* inFilePath, const char* type,
 
     if (strcmp(type, "xml") == 0) {
       ret = cie_sign_sign_set(ctx, CIE_SIGN_OPT_INPUTFILE_TYPE,
-                              reinterpret_cast<void*>(CIE_SIGN_FILETYPE_XML));
+                              reinterpret_cast<void*>(static_cast<uintptr_t>(
+                                  CIE_SIGN_FILETYPE_XML)));
       if (ret != 0) {
         throw ret;
       }
@@ -116,7 +117,8 @@ uint16_t CIESign::sign(const char* inFilePath, const char* type,
       }
 
       ret = cie_sign_sign_set(ctx, CIE_SIGN_OPT_INPUTFILE_TYPE,
-                              reinterpret_cast<void*>(CIE_SIGN_FILETYPE_PDF));
+                              reinterpret_cast<void*>(static_cast<uintptr_t>(
+                                  CIE_SIGN_FILETYPE_PDF)));
       if (ret != 0) {
         throw ret;
       }
